[SECURITY-1956] remove this undocumented feature which has security issue
